added, but not activated smt_disprove rule
comment: the rule is inactive because it makes things slower :(
some very weird profiling results need to be sorted


also

1. Removed Smt.rewrite_star_to_and. Instead, we just tell Z3 that
star is the same as conjunction. As a bonus, we now do so only when
we shipped only pure stuff. (The purity checks should be cached in
the future, because they are now very frequent.)

2. Syntax.is_pure now handles quantifiers.

3. Some catching of Z3native.Exception, to get around what appear to
be rough edges of the Z3 interface. See
  https://z3.codeplex.com/workitem/101
and
  https://z3.codeplex.com/workitem/102

4. Better profiling. Handles functions that raise exceptions.