fixed a bug in Prover.inline_pvars_rule
There were still cases where inline_pvars_rule would incorrectly fire
without making progress (good thing we have that check in
Prover.solve!).

If we have v = p1 * v = p2 in the formula, then v appears twice but not
progress is achieved by substituting. The previous criterion for
usefulness was that v appears at least twice in the formula. Instead, we
have to make sure that v appears more times in the formula than the
numer of times it appears as v = p.