really take assoc of * into account when matching
The current implementation was broken because we used the binary
on_star. This prevents matching from working correctly.

The function on_big_star has made a reappearance. It's been rewritten
differently because the previous one was buggy (it would only descend
into right arguments of *). Perhaps it should use unfold instead but I
couldn't figure out how.

The find_matches complexity has been restored to its former exponential
glory (in the size of the concrete formula to match). This function
should probably be rewritten not to be exponential, since formulas can
grow very large, which makes this a serious issue.

The only way I can see around this would be to stop allowing ?f patterns
at the top level (which require all *-splittings to be explored, of
which there is an exponential number), or perhaps allow only one, which
would then have to be treated very specially (it gets the frame once the
rest of the pattern has been matched). Maybe there was another intent
that I'm missing and find_matches is not supposed to be of exponential
complexity in the first place. Thoughts?