handle emp as the neutral element of * when matching
This is to match a * b against a non-* formula c: rewrite c as c * emp
and continue.

It is possible that this breaks some invariant of find_matches as we
change the formula in the middle of execution. I don't think so because
this triggers only when we are still matching * on the pattern, so
around the beginning.

This would not be needed if we did not require patterns of the form
a * ?f and made sure that the pattern does not contain some * emp.