memoise Smt.check_sat instead of Prover.smt_is_valid
The caching of Prover.smt_is_valid did not take context into account
(pushed assertions) which is an issue for doing more interesting things
with Z3 in Prover.

The caching in Smt takes the context into account and as a result is
slower than the original caching in Prover.smt_is_valid.