updated a couple of tests to new syntax (done with Rasmus)
Also, the mly uses some hackery to be less annoying about parentheses.
There is still a reduce/reduce conflict in the mly. (I *think* I didn't
introduce it.) In any case, the proper fix would be to move abandon
the formula/term distinction in the grammar, and add a proper type-checker
(possibly with inference).